+++ app/Models/User.php
@@ -32,6 +32,8 @@ class User extends Authenticatable
     protected $fillable = [
         'name',
         'email',
+        'phone',
+        'data',
         'password',
     ];
 
@@ -66,6 +68,7 @@ class User extends Authenticatable
         return [
             'email_verified_at' => 'datetime',
             'password' => 'hashed',
+            'data' => 'array',
         ];
     }
 }